+++ src/screens/WalletScreen/Modals/NotAbleToSaveModal/NotAbleToSaveModal.types.ts
